+++ src/cards/MultiTeam/soccer-live-multi-team-editor.js
@@ -1,9 +1,8 @@
 import { LitElement, html, css } from 'lit';
 import { renderSkinControls } from '../skin-editor.js';
 import { t, resolveLang } from '../../i18n.js';
-import { editorStyles } from '../editor-helper.js';
+import { editorStyles, renderLanguageControl } from '../editor-helper.js';
 
-const LANGS = ['auto', 'en', 'nl', 'de', 'pt', 'fr', 'es', 'it'];
 
 class SoccerLiveMultiTeamEditor extends LitElement {
   static get properties() { return { _config: { type: Object }, hass: { type: Object }, _allEntities: { type: Array } }; }
@@ -89,10 +88,7 @@ class SoccerLiveMultiTeamEditor extends LitElement {
           ${renderSkinControls(this, this._config, (k) => (this._t ? this._t(k) : k))}
         </div>
         <div>
-          <label class="field-label">${this._t('editor.language')}</label>
-          <select data-config-value="language" @change=${this._selectChanged}>
-            ${LANGS.map(l => html`<option value="${l === 'auto' ? '' : l}" ?selected=${(this._config.language || '') === (l === 'auto' ? '' : l)}>${l}</option>`)}
-          </select>
+          ${renderLanguageControl(this, this._config, (k) => (this._t ? this._t(k) : k))}
         </div>
       </div>
     `;